requirements.txt
redis
httpx
fastapi
pyodbc
fastapi
uvicorn
llama-cpp-python 
requests